added accumulator for nnue

+++ samaritan/tests/dense.cc
@@ -6,14 +6,14 @@ TEST(TestDense, Forward)
     constexpr std::size_t inSize = 3;
     constexpr std::size_t outSize = 2;
 
-    DenseLayer<float> dense{inSize, outSize};
+    DenseLayer<int> dense{inSize, outSize};
     
     // Example initialization
     dense.setBiases({10, 20, 0});
     dense.setWeights({1, 2, 3, 4, 5, 6});
 
-    std::vector<float> input({1, 2});
-    std::vector<float> output = dense.forward(input);
+    std::vector<int> input({1, 2});
+    std::vector<int> output = dense.forward(input);
 
     EXPECT_EQ(output[0], 15);
     EXPECT_EQ(output[1], 31);
